Only include list in DOM when popover is open
Transcluding the menu items and including them in the DOM only when the popover
is open helps keep the number of watches to a minimum even when views contain
large numbers of popovers menus with numerous bindings on their elements. In
many cases it is likely to also allow one-time bindings to be used to show and
hide menu items assuming that their visibility doesn't need to change while the
menu is open.

Type hints are currently required when calling the $translude function as the
DefinitelyTyped definition for `ITranscludeFunction` has yet to be updated to
include the new parameters for multi-slot transclusion. I have submitted a PR
request for this change [1].

[1] https://github.com/DefinitelyTyped/DefinitelyTyped/pull/10668