Add keyboard navigation
Pressing the Up or Down keys when the popover is closed causes it to open (but
leaves the focus on the button).

Pressing the Up or Down keys when the popover is open causes the focus to move
up or down an item, wrapping round if the top or bottom of the menu is reached.
If the focus is currently on the button, moving 'down' means setting focus on
the first item and moving 'up' means setting focus on the last item.

Pressing the Escape key always closes the popover.

The popover can also be toggled using the standard keys for toggling a button
(although that behaviour is provided by the native button control
implementation, not the code added in this commit).