Add directive for popover menu separators
This directive simply adds the relevant class and role attribute for a separator
item in a popover menu.

The CSS rules take care of hiding separators that are at the top or bottom of
the list along with those that are follow immediately after another separator.
This ensures that `ng-if` bindings can be used to show and hide items within a
popover menu without having to also worry about adding bindings to show and hide
the separators between them.