Close popover menus when their items are clicked
Clicking inside a popover shouldn't normally cause it to close as that might
prevent users from interacting with some of its contents.

However, in the case of a popover menu, it does make sense to close the popover
when the user clicks one of the links within it (but not when the user clicks on
other items such as separators).