Remove wrapper element from popover menus
Previously the `sf-popover-menu` directive would be applied to a wrapper element
containing both the button and the list.

However, now that the popover list is transcluded as a child of the document
body, that wrapper doesn't do anything other than wrap the button.

This commit updates the directive so that it is now applied directly to the
button that toggles the popover, with the list template placed inside the button
tag for transclusion.