Enable two-way controller binding on `isOpen`
Because Angular applies controller bindings prior to the linking phase, this
change was a little more involved than it might otherwise have been.

In particular, when the `isOpen` property setter is first called with its
initial value, the controller's `$postLink` callback has yet to be invoked and
so the usual DOM modifications cannot be applied.

To get around this problem the 'is open' state is now stored in an `_isOpen`
instance variable instead of being deduced from the current state of the DOM. If
the `$postLink` callback has yet to be invoked (which we can determine by
checking whether the `button` instance variable has been initialized) the
`_isOpen` instance variable is updated without any changes being applied to the
DOM (they are applied later by the `$postLink` function instead).